部分注释，int改为long

+++ LCMS/Client/src/main/java/org/businesslogicservice/manageblservice/AccountManagementBLService.java
@@ -1,5 +1,6 @@
 package org.businesslogicservice.manageblservice;
 
+import java.rmi.RemoteException;
 import java.util.ArrayList;
 import java.util.Vector;
 
@@ -9,19 +10,68 @@ import org.po.UserPO;
 import org.vo.BankAccountVO;
 
 public interface AccountManagementBLService {
+	/**
+	 * 获取账户列表VO
+	 * 
+	 * @return Vector<BankAccountVO>
+	 */
 	public Vector<BankAccountVO> getAccountList ();
 	
+	/**
+	 * 添加账户
+	 * 
+	 * @param String
+	 * @return ResultMessage
+	 */
 	public ResultMessage addAccount(String name);
 	
+	/**
+	 * 获取账户余额
+	 * 
+	 * @param String
+	 * @return int
+	 */
 	public int getBalance(String name);
 	
+	/**
+	 * 删除账号
+	 * 
+	 * @param String
+	 * @return ResultMessage
+	 */
 	public ResultMessage delAccount(String name);
 	
+	/**
+	 * 修改账户名
+	 * 
+	 * @param String
+	 * @param String
+	 * @return ResultMessage
+	 */
 	public ResultMessage changeName(String name,String newname);
 	
-	public ResultMessage changeBalance (String name,int change);
+	/**
+	 * 修改余额
+	 * 
+	 * @param String
+	 * @param long
+	 * @return ResultMessage
+	 */
+	public ResultMessage changeBalance (String name,long change);
 	
+	/**
+	 * 关键字查找
+	 * 
+	 * @param String
+	 * @return Vector<BankAccountVO>
+	 */
 	public Vector<BankAccountVO> accountSearch(String namepart);
 	
+	/**
+	 * 判断用户是否有财务人员最高权限
+	 * 
+	 * @param UserPO
+	 * @return boolean
+	 */
 	public boolean check (UserPO user);
 }
